requirements.txt
asgiref==3.8.1
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.8
defusedxml==0.8.0rc2
diff-match-patch==20230430
distlib==0.3.8
Django==4.2.7
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-import-export==4.0.7
django-rest-knox==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
et-xmlfile==1.1.0
fastjsonschema==2.20.0
filelock==3.15.4
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.3.0
msgpack==1.0.8
oauthlib==3.2.2
openpyxl==3.1.3
packaging==24.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg==3.1.19
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.8.0
pyproject_hooks==1.1.0
python3-openid==3.2.0
pywin32-ctypes==0.2.2
rapidfuzz==3.9.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
shellingham==1.5.4
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tablib==3.5.0
tomlkit==0.12.5
trove-classifiers==2024.5.22
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
